Create Forward_pricing_model.py
P(T* + T) =   P(T*)  *   F(T*,T)

The forward contract price of F(1,2) = 0.8262 is the price, 
agreed on today, 
that would be paid one year from today for a bond with a two-year maturity
and 
a risk-free unit- principal payment (e.g., $1, €1, or £1) at maturity. 

As shown in the solution to 3, 
it is calculated as the three-year discount factor, P(3) = 0.7722, divided by the one-year discount factor, P(1) = 0.9346.